refactor(cli-helper, ochat): separate config read and write into the `cli-helper` package
