+++ packages/cli-helper/README_zh.md
@@ -1,0 +1,17 @@
+# @ombro/cli-helper
+
+[![npm package](https://badgen.net/npm/v/@ombro/cli-helper)](https://npmjs.com/package/@ombro/cli-helper)
+
+[English](./README.md) | 简体中文
+
+TODO
+
+## 安装
+
+```sh
+$ npm install @ombro/cli-helper
+```
+
+## 使用
+
+TODO
